Crate jscjs_sys

This crate contains Rust bindings to the Webkit JavaScript engine, JavaScriptCore, developed by Apple.

These bindings are designed to be a fairly straightforward translation to the low-level C API, while taking advantage of Rust’s memory safety. For more about the JavaScriptCore API, see the API source and the documentation.

Provided below are some practical examples of what functionalities this crate allows:

Create a global scripting context, used to create and execute JavaScript objects and code

Work natively with objects, parameters

Build JavaScript functions out of strings

Associate C callbacks to user-definted “classes” of objects

Attach C callbacks to “classes”, handles responses to an action (e.g., getters/setters, promises, fn cals)

Load JavaScript files based on designated names and starting line-numbers

Modules

api

runtime