Function jscjs_sys::api::JSCheckScriptSyntax

pub unsafe extern "C" fn JSCheckScriptSyntax(
    ctx: JSContextRef, 
    script: JSStringRef, 
    sourceURL: JSStringRef, 
    startingLineNumber: c_int, 
    exception: *mut JSValueRef
) -> bool

@function JSCheckScriptSyntax @abstract Checks for syntax errors in a string of JavaScript. @param ctx The execution context to use. @param script A JSString containing the script to check for syntax errors. @param sourceURL A JSString containing a URL for the script’s source file. This is only used when reporting exceptions. Pass NULL if you do not care to include source file information in exceptions. @param startingLineNumber An integer value specifying the script’s starting line number in the file located at sourceURL. This is only used when reporting exceptions. The value is one-based, so the first line is line 1 and invalid values are clamped to 1. @param exception A pointer to a JSValueRef in which to store a syntax error exception, if any. Pass NULL if you do not care to store a syntax error exception. @result true if the script is syntactically correct, otherwise false.